+++ app/build.gradle
@@ -153,8 +153,6 @@ dependencies {
     //implementation 'com.izikode.izilib:accordionrecycler:0.5'
 
 
-
-
     implementation 'com.github.tiper:MaterialSpinner:1.4.0'
 
     implementation 'com.daimajia.easing:library:2.1@aar'
