+++ app/src/main/java/com/sugarcoach/ui/main/presenter/MainPresenterImp.kt
@@ -10,6 +10,7 @@ import io.reactivex.Single
 
 
 interface MainPresenterImp<V : MainView, I : MainInteractorImp> : Presenter<V, I>{
+    fun facebookLogOut()
     fun goToActivityRegister()
     fun goToActivityTreatment()
     fun goToActivityDaily()
