+++ app/src/main/java/com/sugarcoach/ui/main/view/MainView.kt
@@ -6,7 +6,6 @@ import java.util.*
 
 
 interface MainView : BaseView {
-
     fun showSuccessToast()
     fun openRegisterActivity()
     fun openTreatmentActivity()
