+++ app/src/main/java/com/sugarcoach/ui/onboarding/view/OnBoardingActivity.kt
@@ -9,11 +9,13 @@ import androidx.recyclerview.widget.RecyclerView
 import androidx.viewpager.widget.PagerAdapter
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import androidx.viewpager2.widget.ViewPager2
+import com.facebook.AccessToken
 import com.jem.liquidswipe.clippathprovider.LiquidSwipeClipPathProvider
 import com.mazenrashed.dotsindicator.DotsIndicator
 import com.sugarcoach.databinding.ActivityOnboardingBinding
 import com.sugarcoach.ui.base.view.BaseActivity
 import com.sugarcoach.ui.login.view.LoginActivity
+import com.sugarcoach.ui.main.view.MainActivity
 import com.sugarcoach.ui.onboarding.interactor.OnBoardingInteractorImp
 import com.sugarcoach.ui.onboarding.presenter.OnBoardingPresenterImp
 import javax.inject.Inject
@@ -30,6 +32,7 @@ class OnBoardingActivity: BaseActivity(), OnBoardingView {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityOnboardingBinding.inflate(layoutInflater)
+        verifyLogin()
         setContentView(binding.root)
         presenter.onAttach(this)
     }
@@ -51,6 +54,20 @@ class OnBoardingActivity: BaseActivity(), OnBoardingView {
         binding.dotsIndicator.attachTo(binding.boardingPager)
         adapter.setData(itemList)
         adapter.notifyDataSetChanged()
+    }
+
+    override fun verifyLogin() {
+        val accessToken: AccessToken? = AccessToken.getCurrentAccessToken()
+        if(accessToken != null && !accessToken.isExpired){
+            Log.i("OnConfigure", "Usurio loggeado previamente")
+            startMain()
+        }
+    }
+
+    override fun startMain() {
+        val intent = Intent(this, MainActivity::class.java)
+        startActivity(intent)
+        finish()
     }
 
     override fun showProgress() {
