+++ app/src/main/java/com/sugarcoach/ui/onboarding/view/OnBoardingView.kt
@@ -8,4 +8,7 @@ interface OnBoardingView : BaseView {
     fun openLoginActivity()
 
     fun setData(itemList: ArrayList<BoardingItem>)
+    fun verifyLogin()
+
+    fun startMain()
 }
