+++ local.properties
@@ -4,5 +4,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Sun Oct 08 18:13:31 ART 2023
-sdk.dir=C\:\\Users\\ramir\\AppData\\Local\\Android\\Sdk
+#Mon Oct 09 08:34:53 ART 2023
+sdk.dir=C\:\\Users\\PC\\AppData\\Local\\Android\\Sdk
